@dzbarsky 2021-07-10
I didn't see any benchmarks in this repo, how are you testing performance?